Python: catch narrower exception types
Catch `Exception`, instead of `BaseException` or using bare `except:`
clauses.

This is still very broad, but avoids the case of catching `SystemExit`
or `KeyboardInterrupt` (which are not subclasses of `Exception`) and
re-raising them as `ConnectionError`.

Application code may handle `ConnectionError`s differently (such as
attempting re-connection) to `SystemExit` or `KeyboardInterrupt`, which
may be raised in signal handlers and which the application may treat as
a relatively normal shutdown.